requirements.txt
django==3.2.15
django-debug-toolbar==3.2.1
Pillow==8.2.0
environs[django]==9.3.2
djangorestframework==3.13.1

phonenumbers==8.12.55
